// src/schemas/basicformSchema.ts
import { z } from 'zod';

// Define the schema for the Candidate model
export const basicformSchema = z.object({
    firstName: z.string().min(1, "First name must be at least 1 character.").max(255, "First name must be less than 255 characters."),
    middleName: z.string().optional(),
    lastName: z.string().min(1, "Last name must be at least 1 character.").max(255, "Last name must be less than 255 characters."),
    nametype: z.string(),
    dob: z.date(),
    passportNumber: z.string().length(8, { message: "Enter valid passport number" }),
    passportType: z.string(),
    passportCountry: z.string(),
    ssn: z.string().length(9, { message: "Enter valid SSN" }),
    address: z.string().min(1, "Address must be at least 1 character.").max(255, "Address must be less than 255 characters."),
    city: z.string(),
    state: z.string().min(1, "State must be at least 1 character."),
    zip: z.string().min(1, "Enter a valid number"),
    jobCity: z.string().optional(),
    jobState: z.string().optional(),
    jobZip: z.string().optional(),
    positionTitle: z.string().optional(),
    applicantPhone: z.string().min(10, { message: 'Must be a valid mobile number' }).max(14, { message: 'Must be a valid mobile number' }),
    applicantEmail: z.string().email("Invalid email address."),
    race: z.string().optional(),
    gender: z.string().optional(),
    dlCountry: z.string(),
    dlNumber: z.string().min(6, "Number must be at least 6 characters.").max(20, "Number must be less than 20 characters."),
    dlState: z.string(),
    citizenship: z.string(),
    motherMaidenName: z.string(),
    motherFullName: z.string(),
    fatherName: z.string(),
    cityOfBirth: z.string(),
    countryOfBirth: z.string(),
    fcraPurpose: z.string(),
    statedMonthlyIncome: z.string(),
    proposedMonthlyRent: z.string(),
    knownCriminalRecords: z.string(),
    criminalRecordDescription: z.string().optional(),
});